Test using a json parser
Testing using string compares way to fragile and asking for trouble.
Also, we're only really interested in the value(s) of the resulting json
object, what data it parses to. Whether or not they add a space after
every comma or not is not relevant to us but it is for string
comparison.