Doughnuts, plugins, title, legend and other data points types
Major refactoring of the ChartJs configuration to change the K-V types
to a single E (element) type. This allows for simple arrays of data as
is needed by a doughnut or pie chart.

Configuration of the title, legend and plugins was added.

The markup of SimpleGraphPanel was reduced to the minimum required to
render a chart.

SimpleGraphPanel now supports updates via Ajax. It is not yet possible
to add or remove datasets.