temp fix for bug during javadoc generation
using javadoc on jdk 11 with source 8 fails with an error about using
modules but being on the unnamed module path. This should have been
fixed in openjdk 11.0.3 but it seems to still happen.

Setting the 'source' configuration option to 8 for the javadoc plugin is
a work around to keep the project building.

See JDK-8212233